Add participation filter for participations without submissions + add submission count column to participations table
