+++ src/main/webapp/app/entities/participation/participation.component.html
@@ -13,6 +13,15 @@
                     <input type="radio" [ngModel]="participationCriteria.filterProp" (click)="updateParticipationFilter(FilterProp.FAILED)" [value]="FilterProp.FAILED" />
                     <span class="ml-1">{{ 'artemisApp.exercise.showFailed' | translate }}</span>
                 </label>
+                <label class="radio-inline ml-2 mb-0 d-flex align-items-center">
+                    <input
+                        type="radio"
+                        [ngModel]="participationCriteria.filterProp"
+                        (click)="updateParticipationFilter(FilterProp.NO_SUBMISSIONS)"
+                        [value]="FilterProp.NO_SUBMISSIONS"
+                    />
+                    <span class="ml-1">{{ 'artemisApp.exercise.showNoSubmissions' | translate }}</span>
+                </label>
             </div>
         </div>
         <jhi-programming-exercise-instructor-submission-state
@@ -127,6 +136,27 @@
                     </ng-template>
                     <ng-template ngx-datatable-cell-template let-value="value">
                         {{ value | date: 'medium' }}
+                    </ng-template>
+                </ngx-datatable-column>
+
+                <ngx-datatable-column [minWidth]="100" [width]="110">
+                    <ng-template ngx-datatable-header-template>
+                        <span class="datatable-header-cell-wrapper" (click)="controls.onSort('submissionCount')">
+                            <span class="datatable-header-cell-label bold sortable" jhiTranslate="artemisApp.exercise.submissionCount">
+                                Submissions
+                            </span>
+                            <fa-icon [icon]="controls.iconForSortPropField('submissionCount')"></fa-icon>
+                        </span>
+                    </ng-template>
+                    <ng-template ngx-datatable-cell-template let-value="row">
+                        <div class="text-center">
+                            <div *ngIf="exercise.isAtLeastInstructor; else displayTextOnly">
+                                <a routerLink="/participation/{{ value.id }}/submissions">{{ value.submissionCount }}</a>
+                            </div>
+                            <ng-template #displayTextOnly>
+                                {{ value.submissionCount }}
+                            </ng-template>
+                        </div>
                     </ng-template>
                 </ngx-datatable-column>
