+++ src/main/webapp/app/scores/exercise-scores.component.html
@@ -159,7 +159,7 @@
                     </ng-template>
                 </ngx-datatable-column>
 
-                <ngx-datatable-column prop="submissionCount" [resizeable]="false" [sortable]="true" [minWidth]="110" [width]="120">
+                <ngx-datatable-column prop="participation" [resizeable]="false" [sortable]="true" [minWidth]="110" [width]="120">
                     <ng-template ngx-datatable-header-template>
                         <span class="datatable-header-cell-wrapper" (click)="controls.onSort('submissionCount')">
                             <span class="datatable-header-cell-label bold">{{ 'artemisApp.exercise.submissionCount' | translate }}</span>
@@ -167,7 +167,14 @@
                         </span>
                     </ng-template>
                     <ng-template ngx-datatable-cell-template let-value="value">
-                        {{ value }}
+                        <div class="text-center">
+                            <div *ngIf="exercise.isAtLeastInstructor; else displayTextOnly">
+                                <a routerLink="/participation/{{ value.id }}/submissions">{{ value.submissionCount }}</a>
+                            </div>
+                            <ng-template #displayTextOnly>
+                                {{ value.submissionCount }}
+                            </ng-template>
+                        </div>
                     </ng-template>
                 </ngx-datatable-column>
