+++ src/main/webapp/i18n/en/exercise.json
@@ -64,6 +64,7 @@
             "duration": "Duration",
             "showBuildFailed": "Show Build Failed",
             "showFailed": "Show Failed Submissions (without Result)",
+            "showNoSubmissions": "Show No Submissions",
             "showUnsuccessful": "Show Unsuccessful",
             "showSuccessful": "Show Successful",
             "showManual": "Show Manual",
